requirements.txt
langchain-core>=0.3.51
langchain-openai>=0.0.5
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.3
numpy>=1.26.3
pandas>=2.1.4
setuptools==67.6.1
langchain>=0.3.24